refactor: dto 구조 변경
