+++ src/main/java/Bob_BE/domain/discount/dto/request/DiscountRequestDto.java
@@ -1,7 +1,5 @@
 package Bob_BE.domain.discount.dto.request;
 
-import Bob_BE.domain.discount.dto.data.DiscountDataDto;
-import Bob_BE.domain.discountMenu.dto.data.DiscountMenuDataDto;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +7,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
 public class DiscountRequestDto {
@@ -20,12 +17,22 @@ public class DiscountRequestDto {
     @AllArgsConstructor
     public static class CreateDiscountRequestDto {
 
-        private List<DiscountMenuDataDto.CreateDiscountMenuDataDto> discountMenuDataDtoList;
+        private List<CreateDiscountMenuDataDto> discountMenuDataDtoList;
         @NotNull
         private String title;
         @NotNull
         private LocalDate startDate;
         @NotNull
         private LocalDate endDate;
+    }
+
+    @Getter
+    @Builder
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class CreateDiscountMenuDataDto {
+
+        private Long menuId;
+        private Integer discountPrice;
     }
 }
